Fixed floating-point division issue

+++ helper.py
@@ -150,7 +150,7 @@ def compose_bookname(title, author, edition, isbn, max_length):
 def get_random_string(length):
     sha512 = hashlib.sha512(str(time.time()).encode('utf-8'))
     name = ''
-    for i in range(0, (length / 128 + 1)):
+    for i in range(0, int(length / 128 + 1)):
         sha512.update(str(time.time()).encode('utf-8'))
         name = name + sha512.hexdigest()
     return name[:length]
@@ -172,5 +172,5 @@ def get_max_filename_length(path):
                 name = get_random_string(mid)
                 continue
             hi = mid
-        mid = (hi + lo) / 2
+        mid = int((hi + lo) / 2)
     return mid
